+++ lib/iterable.dart
@@ -2,6 +2,7 @@
 export 'src/iterable/chunked.dart' show ChunkedExtension;
 export 'src/iterable/combinations.dart' show CombinationsExtension;
 export 'src/iterable/cycle.dart' show CycleExtension;
+export 'src/iterable/entries.dart' show EntriesExtension;
 export 'src/iterable/flat_map.dart' show FlatMapExtension;
 export 'src/iterable/flatten.dart' show FlattenExtension;
 export 'src/iterable/group.dart' show GroupExtension, Group;
@@ -15,6 +16,8 @@ export 'src/iterable/product.dart' show ProductExtension;
 export 'src/iterable/random.dart' show RandomExtension;
 export 'src/iterable/repeat.dart' show RepeatExtension, repeat;
 export 'src/iterable/separated.dart' show SeparatedExtension;
+export 'src/iterable/to_map.dart'
+    show ToMapIterableExtension, ToMapIterableMapEntryExtension;
 export 'src/iterable/unique.dart' show UniqueExtension;
 export 'src/iterable/window.dart' show WindowExtension;
 export 'src/iterable/zip.dart' show ZipExtension;